Move most CPCe file upload/export functionality to a new `cpce` app.
