+++ project/cpce/static/js/UploadAnnotationsCPCHelper.js
