+++ project/export/urls.py
@@ -6,12 +6,6 @@ urlpatterns = [
          views.export_metadata, name="export_metadata"),
     path('annotations/',
          views.export_annotations, name="export_annotations"),
-    path('annotations_cpc_create_ajax/',
-         views.export_annotations_cpc_create_ajax,
-         name="export_annotations_cpc_create_ajax"),
-    path('annotations_cpc_serve/',
-         views.export_annotations_cpc_serve,
-         name="export_annotations_cpc_serve"),
     path('image_covers/',
          views.ImageCoversExportView.as_view(), name="export_image_covers"),
     path('labelset/',
