+++ project/export/utils.py
@@ -1,14 +1,10 @@
 import csv
-from io import StringIO
-from pathlib import PureWindowsPath
 from zipfile import ZipFile
 
 from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.http import HttpResponse
 
-from annotations.model_utils import AnnotationAreaUtils
-from images.models import Point
 from images.utils import metadata_field_names_to_labels
 from vision_backend.utils import get_label_scores_for_point
 from visualization.forms import create_image_filter_form
@@ -56,298 +52,6 @@ def create_csv_stream_response(filename):
 def create_zip_stream_response(filename):
     # https://stackoverflow.com/a/29539722/
     return create_stream_response('application/zip', filename)
-
-
-def create_cpc_strings(image_set, cpc_prefs):
-    # Dict mapping from cpc filenames to cpc file contents as strings.
-    cpc_strings = dict()
-
-    for img in image_set:
-        # Write .cpc contents to a stream.
-        cpc_stream = StringIO()
-
-        if img.cpc_content and img.cpc_filename:
-            # A CPC file was uploaded for this image before.
-            write_annotations_cpc_based_on_prev_cpc(cpc_stream, img, cpc_prefs)
-            # Use the same CPC filename that was used for this image before.
-            cpc_filename = img.cpc_filename
-        else:
-            # No CPC file was uploaded for this image before.
-            write_annotations_cpc(cpc_stream, img, cpc_prefs)
-            # Make a CPC filename based on the image filename, like CPCe does.
-            # PWP ensures that both forward slashes and backslashes are counted
-            # as path separators.
-            cpc_filename = image_filename_to_cpc_filename(
-                PureWindowsPath(img.metadata.name).name)
-
-        # If the image name seems to be a relative path (not just a filename),
-        # then use those path directories on the CPC .zip filepath as well.
-        image_parent = PureWindowsPath(img.metadata.name).parent
-        # If it's a relative path, this appends the directories, else this
-        # appends nothing.
-        cpc_filepath = str(PureWindowsPath(image_parent, cpc_filename))
-        # We've used Windows paths for path-separator flexibility up to this
-        # point. Now that we're finished with path manipulations, we'll make
-        # sure all separators are forward slashes for .zip export purposes.
-        # This makes zip directory tree structures work on every OS. Forward
-        # slashes are also required by the .ZIP File Format Specification.
-        # https://superuser.com/a/1382853/
-        cpc_filepath = cpc_filepath.replace('\\', '/')
-
-        # Convert the stream contents to a string.
-        # TODO: If cpc_filepath is already in the cpc_strings dict, then we
-        # have a name conflict and need to warn / disambiguate.
-        cpc_strings[cpc_filepath] = cpc_stream.getvalue()
-
-    return cpc_strings
-
-
-def create_zipped_cpcs_stream_response(cpc_strings, zip_filename):
-    response = create_zip_stream_response(zip_filename)
-    # Convert Unicode strings to byte strings
-    cpc_byte_strings = dict([
-        (cpc_filename, cpc_content.encode())
-        for cpc_filename, cpc_content in cpc_strings.items()
-    ])
-    write_zip(response, cpc_byte_strings)
-    return response
-
-
-def image_filename_to_cpc_filename(image_filename):
-    """
-    Take an image filename string and convert to a cpc filename according
-    to CPCe's rules. As far as we can tell, it's simple: strip extension,
-    add '.cpc'. Examples:
-    IMG_0001.JPG -> IMG_0001.cpc
-    img 0001.jpg -> img 0001.cpc
-    my_image.bmp -> my_image.cpc
-    another_image.gif -> another_image.cpc
-    """
-    cpc_filename = PureWindowsPath(image_filename).stem + '.cpc'
-    return cpc_filename
-
-
-def point_to_cpc_export_label_code(point, annotation_filter=None):
-    """
-    Normally, annotation export will export ALL annotations, including machine
-    annotations of low confidence. This is usually okay because, if users want
-    to exclude Unconfirmed annotations when exporting, they can filter images
-    to just Confirmed and then export from there.
-
-    However, this breaks down in CPC export's use case: CPC import,
-    add confident machine annotations, and then CPC export to continue
-    annotating in CPCe. These users will expect to only export the confident
-    machine annotations, and confidence is on a per-point basis, not per-image.
-    So we need to filter the annotations on a point basis. That's what this
-    function is for.
-
-    :param point: A Point model object.
-    :param annotation_filter:
-      'confirmed_only' to denote that only Confirmed annotations are accepted.
-      'confirmed_and_confident' to denote that Unconfirmed annotations above
-      the source's confidence threshold are also accepted. Normally, these
-      annotations will become Confirmed when you enter the annotation tool for
-      that image... but if you're planning to annotate in CPCe, there's no
-      other reason to enter the annotation tool!
-    :return: Label short code string of the point's annotation, if there is
-      an annotation which is accepted by the annotation_filter. Otherwise, ''.
-    """
-    if annotation_filter is None:
-        return point.label_code
-
-    if point.annotation_status_property == 'confirmed':
-        # Confirmed annotations are always included
-        return point.label_code
-    elif (annotation_filter == 'confirmed_and_confident'
-          and point.annotation_status_property == 'unconfirmed'):
-        # With this annotation_filter, Unconfirmed annotations are included
-        # IF they're above the source's confidence threshold
-        if point.machine_confidence >= point.image.source.confidence_threshold:
-            return point.label_code
-
-    # The annotation filter rejects this annotation, or there is no annotation
-    return ''
-
-
-def write_annotations_cpc(cpc_stream, img, cpc_prefs):
-    """
-    Write a CPC from scratch.
-    :param cpc_stream: An IO stream object.
-    :param img: An Image model object.
-    :param cpc_prefs:
-      CPC 'preferences' indicated by previous CPC uploads to this source.
-      Includes code file path and image directory.
-      These are not important to usage of the image within CoralNet, but they
-      are important for exporting back to CPC as seamlessly as possible.
-    :return:
-    """
-    # Each line is a series of comma-separated tokens, so the CSV module works
-    # well enough.
-    # Some tokens are quoted and some aren't, seemingly without checking for
-    # presence of spaces, etc. We'll handle quotes on our own to ensure we
-    # match this behavior.
-    writer = csv.writer(cpc_stream, quotechar=None, quoting=csv.QUOTE_NONE)
-
-    # Line 1: Environment info and image dimensions
-    local_image_path = PureWindowsPath(
-        cpc_prefs['local_image_dir'], img.metadata.name)
-    row = [
-        '"' + cpc_prefs['local_code_filepath'] + '"',
-        '"' + str(local_image_path) + '"',
-        # Image dimensions. CPCe typically operates in units of 1/15th of a
-        # pixel. If a different DPI setting is used in an older version of CPCe
-        # (like CPCe 3.5), it can be something else like 1/12th. But we'll
-        # assume 1/15th for export.
-        img.original_width * 15,
-        img.original_height * 15,
-        # These 2 items seem to be the display width/height that the image
-        # was opened at in CPCe. Last opened? Initially opened? Not sure.
-        # Is this ever even used when opening the CPC file? As far as we know,
-        # no. We'll just arbitrarily set this to 960x720.
-        # The CPCe documentation says CPCe works best with 1024x768 resolution
-        # and above, so displaying the image itself at 960x720 is roughly
-        # around there.
-        960 * 15,
-        720 * 15,
-    ]
-    writer.writerow(row)
-
-    # Lines 2-5: Annotation area bounds.
-    # <x from left>,<y from top> in units of 1/15th of a pixel.
-    # Order: Bottom left, bottom right, top right, top left.
-    # Get from the image model if present, otherwise make it the whole image.
-
-    anno_area = AnnotationAreaUtils.db_format_to_numbers(
-        img.metadata.annotation_area)
-    anno_area_type = anno_area.pop('type')
-    if anno_area_type == AnnotationAreaUtils.TYPE_PIXELS:
-        # This is the format we want
-        pass
-    elif anno_area_type == AnnotationAreaUtils.TYPE_PERCENTAGES:
-        # Convert to pixels
-        anno_area = AnnotationAreaUtils.percentages_to_pixels(
-            width=img.original_width, height=img.original_height, **anno_area)
-    elif anno_area_type == AnnotationAreaUtils.TYPE_IMPORTED:
-        # Unspecified; just use the whole image
-        anno_area = dict(
-            min_x=0, max_x=img.max_column,
-            min_y=0, max_y=img.max_row)
-    bound_left = anno_area['min_x'] * 15
-    bound_right = anno_area['max_x'] * 15
-    bound_top = anno_area['min_y'] * 15
-    bound_bottom = anno_area['max_y'] * 15
-    writer.writerow([bound_left, bound_bottom])
-    writer.writerow([bound_right, bound_bottom])
-    writer.writerow([bound_right, bound_top])
-    writer.writerow([bound_left, bound_top])
-
-    # Line 6: number of points
-    points = Point.objects.filter(image=img).order_by('point_number')
-    writer.writerow([points.count()])
-
-    # Next num_points lines: point positions.
-    # <x from left, y from top> of each point in numerical order,
-    # seemingly using the x15 scaling.
-    # CPCe point positions are on a scale of 15 units = 1 pixel, and
-    # the positions start from 0.
-    for point in points:
-        point_left = point.column * 15
-        point_top = point.row * 15
-        row = [point_left, point_top]
-        writer.writerow(row)
-
-    # Next num_points lines: point labels.
-    # "<point number/letter>","<label code>","Notes","<notes code>"
-    for point in points:
-        label_code = point_to_cpc_export_label_code(
-            point, cpc_prefs['annotation_filter'])
-
-        if cpc_prefs['plus_notes'] and '+' in label_code:
-            # Assumption: label code in CoralNet source's labelset
-            # == {ID}+{Notes} in the .cpc file (case insensitive),
-            # and {ID} does not have a + character in it.
-            cpc_id, cpc_notes = label_code.split('+', maxsplit=1)
-            row = [point.point_number, cpc_id, 'Notes', cpc_notes]
-        else:
-            # Assumption: label code in CoralNet source's labelset
-            # == label code in the .cpc file (case insensitive).
-            row = [point.point_number, label_code, 'Notes', '']
-
-        row = ['"{s}"'.format(s=s) for s in row]
-        writer.writerow(row)
-
-    # Next 28 lines: header fields.
-    for _ in range(28):
-        writer.writerow(['" "'])
-
-
-def write_annotations_cpc_based_on_prev_cpc(cpc_stream, img, cpc_prefs):
-    old_cpc = StringIO(img.cpc_content)
-
-    # Copy first 5 lines
-    for _ in range(5):
-        cpc_stream.write(next(old_cpc))
-
-    # 6th line has the point count
-    point_count_line = next(old_cpc)
-    point_count = int(point_count_line.strip())
-    cpc_stream.write(point_count_line)
-
-    # Next point_count lines have the point positions
-    for _ in range(point_count):
-        cpc_stream.write(next(old_cpc))
-
-    # Next point_count lines have the labels.
-    # Replace the label codes (and notes, if applicable) with the data from
-    # CoralNet's DB.
-    points = Point.objects.filter(image=img).order_by('point_number')
-
-    for point in points:
-        old_line = next(old_cpc)
-        old_tokens = old_line.split(',')
-
-        label_code = point_to_cpc_export_label_code(
-            point, cpc_prefs['annotation_filter'])
-
-        # Change the ID code (and Notes code if applicable), and ensure
-        # they're in double quotes.
-        # The rest of this line stays the same.
-        if cpc_prefs['plus_notes']:
-            # Get Notes from CoralNet's label codes.
-            if '+' in label_code:
-                cpc_id, cpc_notes = label_code.split('+', maxsplit=1)
-            else:
-                cpc_id = label_code
-                cpc_notes = ''
-            # When replacing the last token (notes code), we have to be
-            # careful to preserve the newline.
-            new_tokens = [
-                old_tokens[0], f'"{cpc_id}"',
-                old_tokens[2], f'"{cpc_notes}"\r\n']
-        else:
-            # Do not get Notes from CoralNet's label codes.
-            new_tokens = [
-                old_tokens[0], f'"{label_code}"',
-                old_tokens[2], old_tokens[3]]
-        new_line = ','.join(new_tokens)
-        cpc_stream.write(new_line)
-
-    # Copy remaining contents
-    cpc_stream.write(old_cpc.read())
-
-
-def get_previous_cpcs_status(image_set):
-    if image_set.exclude(cpc_content='').exists():
-        # At least 1 image has a previous CPC
-        if image_set.filter(cpc_content='').exists():
-            # Some, but not all images have previous CPCs
-            return 'some'
-        else:
-            # All images have previous CPCs
-            return 'all'
-    else:
-        # No images have previous CPCs
-        return 'none'
 
 
 def write_zip(zip_stream, file_strings):
