+++ project/lib/templates/lib/nav_test.html
@@ -201,7 +201,7 @@
           <span>Annotations (CSV)</span>
         </a></li>
 
-        <li><a href="{% url 'upload_annotations_cpc' source.pk %}">
+        <li><a href="{% url 'cpce:upload_page' source.pk %}">
           <span>Annotations (CPC)</span>
         </a></li>
