+++ project/upload/utils.py
@@ -1,10 +1,7 @@
 import codecs
 from collections import OrderedDict
 import csv
-import functools
 from io import StringIO
-from pathlib import PureWindowsPath
-import re
 
 # PyCharm may warn that this isn't declared in __all__, but this import
 # simply matches BS4's docs as of 2018/12:
@@ -437,355 +434,6 @@ def annotations_csv_verify_contents(csv_annotations, source):
         raise FileProcessError("No matching image names found in the source")
 
     return annotations
-
-
-def annotations_cpcs_to_dict_read_csv_row(
-        reader, cpc_filename, num_tokens_expected):
-    """
-    Helper for annotations_cpcs_to_dict().
-    Basically like calling next(reader), but with more controlled error
-    handling.
-    :param reader: A CSV reader object.
-    :param cpc_filename: The filename of the .cpc file we're reading.
-     Used for error messages.
-    :param num_tokens_expected: Number of tokens expected in this CSV row.
-    :return: iterable of the CSV tokens, stripped of leading and
-     trailing whitespace.
-    """
-    try:
-        row_tokens = [token.strip() for token in next(reader)]
-    except StopIteration:
-        raise FileProcessError(
-            "File {cpc_filename} seems to have too few lines.".format(
-                cpc_filename=cpc_filename))
-
-    if len(row_tokens) != num_tokens_expected:
-        raise FileProcessError((
-            "File {cpc_filename}, line {line_num} has"
-            " {num_tokens} comma-separated tokens, but"
-            " {num_tokens_expected} were expected.").format(
-                cpc_filename=cpc_filename,
-                line_num=reader.line_num,
-                num_tokens=len(row_tokens),
-                num_tokens_expected=num_tokens_expected))
-
-    return row_tokens
-
-
-def annotations_cpcs_to_dict(cpc_names_and_streams, source, plus_notes):
-    """
-    :param cpc_names_and_streams: An iterable of Coral Point Count .cpc files
-      as (file name, Unicode stream) tuples. One .cpc corresponds to one image.
-    :param source: The source these .cpc files correspond to.
-    :param plus_notes: Boolean option on whether to interpret the + char as
-      an ID/Notes separator.
-    :return: A dict of relevant info, with the following keys::
-      annotations: dict of (image ids -> lists of dicts with keys row,
-        column, (opt.) label).
-      cpc_contents: dict of image ids -> .cpc's full contents as a string
-      code_filepath: Local path to CPCe code file used by one of the .cpc's.
-        Since CPCe is Windows only, this is going to be a Windows path.
-      image_dir: Local path to image directory used by one of the .cpc's.
-        Again, a Windows path. Ending slash can be present or not.
-    Throws a FileProcessError if a .cpc has a problem.
-    """
-
-    cpc_dicts = []
-    code_filepath = None
-
-    for cpc_name, cpc_stream in cpc_names_and_streams:
-
-        cpc_dict = dict(cpc_filename=cpc_name)
-
-        # Each line of a .cpc file is like a CSV row.
-        #
-        # But different lines have different kinds of values, so we'll go
-        # through the lines with next() instead of with a for loop.
-        reader = csv.reader(cpc_stream, delimiter=',', quotechar='"')
-        read_csv_row_curried = functools.partial(
-            annotations_cpcs_to_dict_read_csv_row, reader, cpc_name)
-
-        # Line 1
-        code_filepath, image_filepath, width, height, _, _ = \
-            read_csv_row_curried(6)
-        cpc_dict['image_filepath'] = image_filepath
-        cpc_dict['image_width_in_cpce_scale'] = width
-        cpc_dict['image_height_in_cpce_scale'] = height
-
-        # Lines 2-5; don't need any info from these,
-        # but they should have 2 tokens each
-        for _ in range(4):
-            read_csv_row_curried(2)
-
-        # Line 6: number of points
-        token = read_csv_row_curried(1)[0]
-        try:
-            num_points = int(token)
-            if num_points <= 0:
-                raise ValueError
-        except ValueError:
-            raise FileProcessError((
-                "File {cpc_filename}, line {line_num} is supposed to have"
-                " the number of points, but this line isn't a"
-                " positive integer: {token}").format(
-                    cpc_filename=cpc_name,
-                    line_num=reader.line_num,
-                    token=token))
-
-        # Next num_points lines: point positions. x (column), then y (row).
-        cpc_dict['points'] = []
-        for _ in range(num_points):
-            x_str, y_str = read_csv_row_curried(2)
-            cpc_dict['points'].append(dict(x_str=x_str, y_str=y_str))
-
-        # Next num_points lines: point labels.
-        # We're taking advantage of the fact that the previous section
-        # and this section are both in point-number order. As long as we
-        # maintain that order, we assign labels to the correct points.
-        for point_index in range(num_points):
-            _, cpc_id, _, cpc_notes = read_csv_row_curried(4)
-            if cpc_id:
-                if cpc_notes and plus_notes:
-                    # Assumption: label code in CoralNet source's labelset
-                    # == {ID}+{Notes} in the .cpc file (case insensitive).
-                    cpc_dict['points'][point_index]['label'] = \
-                        f'{cpc_id}+{cpc_notes}'
-                else:
-                    # Assumption: label code in CoralNet source's labelset
-                    # == label code in the .cpc file (case insensitive).
-                    cpc_dict['points'][point_index]['label'] = cpc_id
-
-        cpc_stream.seek(0)
-        cpc_dict['cpc_content'] = cpc_stream.read()
-
-        cpc_dicts.append(cpc_dict)
-
-    # So far we've checked the .cpc formatting. Now check the validity
-    # of the contents.
-    cpc_annotations, cpc_contents, cpc_filenames, image_dir = \
-        annotations_cpc_verify_contents(cpc_dicts, source)
-
-    return dict(
-        annotations=cpc_annotations,
-        cpc_contents=cpc_contents,
-        cpc_filenames=cpc_filenames,
-        code_filepath=code_filepath,
-        image_dir=image_dir,
-    )
-
-
-def annotations_cpc_verify_contents(cpc_dicts, source):
-    """
-    Argument dict is a list of dicts, one dict per .cpc file.
-
-    We'll create a new annotations dict indexed
-    by image id, while verifying image existence, row, column, and label.
-
-    While we're at it, we'll also make a dict from image id to .cpc contents.
-    """
-    annotations = OrderedDict()
-    cpc_contents = OrderedDict()
-    cpc_filenames = OrderedDict()
-    image_names_to_cpc_filenames = dict()
-    image_dir = None
-
-    for cpc_dict in cpc_dicts:
-
-        cpc_filename = cpc_dict['cpc_filename']
-
-        # The image filepath follows the rules of the OS running CPCe,
-        # not the rules of the server OS. So we don't use Path.
-        # CPCe only runs on Windows, so we can assume it's a Windows
-        # path. That means using PureWindowsPath (WindowsPath can only
-        # be instantiated on a Windows OS).
-        cpc_image_filepath = PureWindowsPath(cpc_dict['image_filepath'])
-        image_filename = cpc_image_filepath.name
-
-        # Match up the CPCe image filepath to an image name on CoralNet.
-        #
-        # Let's say the image filepath is D:\Site A\Transect 1\01.jpg
-        # Example image names:
-        # D:\Site A\Transect 1\01.jpg: best match
-        # Site A\Transect 1\01.jpg: 2nd best match
-        # Transect 1\01.jpg: 3rd best match
-        # 01.jpg: 4th best match
-        # Transect 1/01.jpg: same as with backslash
-        # /Transect 1/01.jpg: same as without leading slash
-        # 23.jpg: non-match 1
-        # 4501.jpg: non-match 2
-        # sect 1\01.jpg: non-match 3
-        #
-        # First get names consisting of 01.jpg preceded by /, \, or nothing.
-        # This avoids non-match 1 and non-match 2.
-        regex_escaped_filename = re.escape(image_filename)
-        name_regex = r'^(.*[\\|/])?{fn}$'.format(fn=regex_escaped_filename)
-        image_candidates = source.image_set.filter(
-            metadata__name__regex=name_regex)
-        # Find the best match while avoiding non-match 3.
-        img = None
-        for image_candidate in image_candidates:
-            # Passing through PWP ensures we consistently use backslashes
-            # instead of forward slashes
-            image_name = str(PureWindowsPath(image_candidate.metadata.name))
-            if image_name == str(cpc_image_filepath):
-                # Best possible match
-                img = image_candidate
-                image_dir = ''
-                break
-
-            # Iterate over parents, from top (D:\)
-            # to bottom (D:\Site A\Transect 1).
-            # If a parent combines with image_name to form the full path, then
-            # we have a match.
-            # TODO: This won't work if image_name starts with a slash; do we
-            # want to accommodate that or not?
-            for parent in cpc_image_filepath.parents:
-                if PureWindowsPath(parent, image_name) == cpc_image_filepath:
-                    # It's a match
-                    if img and len(img.metadata.name) > len(image_name):
-                        # There's already a longer match
-                        pass
-                    else:
-                        # This is the best match so far
-                        img = image_candidate
-                        image_dir = str(parent)
-                    # Move onto the next candidate
-                    break
-
-        if img is None:
-            # No matching image names in the source. Just skip this CPC
-            # without raising an error. It could be an image the user is
-            # planning to upload later, or an image they're not planning
-            # to upload but are still tracking in their records.
-            continue
-
-        image_name = img.metadata.name
-
-        if image_name in image_names_to_cpc_filenames:
-            raise FileProcessError((
-                "Image {name} has points from more than one .cpc file: {f1}"
-                " and {f2}. There should be only one .cpc file"
-                " per image.").format(
-                    name=image_name,
-                    f1=image_names_to_cpc_filenames[image_name],
-                    f2=cpc_filename,
-                )
-            )
-        image_names_to_cpc_filenames[image_name] = cpc_filename
-
-        # Detect pixel scale factor - the scale of the x, y units CPCe used to
-        # express the point locations.
-        #
-        # This is normally 15 units per pixel, but
-        # that only holds when CPCe runs in 96 DPI. Earlier versions of CPCe
-        # (such as CPCe 3.5) did not enforce 96 DPI, so for example, it is
-        # possible to run in 120 DPI and get a scale of 12 units per pixel.
-        #
-        # We can figure out the scale factor by reading the .cpc file's image
-        # resolution values. These values are in CPCe's scale, and we know the
-        # resolution in pixels, so we can solve for the scale factor.
-        try:
-            cpce_scale_width = int(cpc_dict['image_width_in_cpce_scale'])
-            cpce_scale_height = int(cpc_dict['image_height_in_cpce_scale'])
-        except ValueError:
-            raise FileProcessError(
-                "File {cpc_filename}: The image width and height on line 1"
-                " must be integers.".format(cpc_filename=cpc_filename))
-
-        x_scale = cpce_scale_width / img.original_width
-        y_scale = cpce_scale_height / img.original_height
-        if (not x_scale.is_integer()
-                or not y_scale.is_integer()
-                or x_scale != y_scale):
-            raise FileProcessError(
-                "File {cpc_filename}: Could not establish an integer scale"
-                " factor from line 1.".format(cpc_filename=cpc_filename))
-        pixel_scale_factor = x_scale
-
-        annotations_for_image = []
-
-        for point_number, cpc_point_dict in enumerate(cpc_dict['points'], 1):
-
-            # Check that row/column are integers within the image dimensions.
-            # Convert from CPCe units to pixels in the meantime.
-
-            point_error_prefix = \
-                "From file {cpc_filename}, point {point_number}:".format(
-                    cpc_filename=cpc_filename, point_number=point_number)
-
-            try:
-                y = int(cpc_point_dict['y_str'])
-                if y < 0:
-                    raise ValueError
-            except ValueError:
-                raise FileProcessError((
-                    point_error_prefix +
-                    " Row should be a non-negative integer, not"
-                    " {y_str}").format(
-                        y_str=cpc_point_dict['y_str']))
-
-            try:
-                x = int(cpc_point_dict['x_str'])
-                if x < 0:
-                    raise ValueError
-            except ValueError:
-                raise FileProcessError((
-                    point_error_prefix +
-                    " Column should be a non-negative integer, not"
-                    " {x_str}").format(
-                        x_str=cpc_point_dict['x_str']))
-
-            # CPCe units -> pixels conversion.
-            row = int(round(y / pixel_scale_factor))
-            column = int(round(x / pixel_scale_factor))
-            point_dict = dict(
-                row=row,
-                column=column,
-            )
-
-            if row > img.max_row:
-                raise FileProcessError(
-                    point_error_prefix +
-                    " Row value of {y} corresponds to pixel {row},"
-                    " but image {name} is only {height} pixels high"
-                    " (accepted values are 0~{max_row})".format(
-                        y=y, row=row, name=image_name,
-                        height=img.original_height,
-                        max_row=img.max_row))
-
-            if column > img.max_column:
-                raise FileProcessError(
-                    point_error_prefix +
-                    " Column value of {x} corresponds to pixel {column},"
-                    " but image {name} is only {width} pixels wide"
-                    " (accepted values are 0~{max_column})".format(
-                        x=x, column=column, name=image_name,
-                        width=img.original_width,
-                        max_column=img.max_column))
-
-            if 'label' in cpc_point_dict:
-                # Check that the label is in the labelset
-                label_code = cpc_point_dict['label']
-                if not source.labelset.get_global_by_code(label_code):
-                    raise FileProcessError(
-                        point_error_prefix +
-                        " No label of code {code} found"
-                        " in this source's labelset".format(
-                            code=label_code))
-                point_dict['label'] = label_code
-
-            annotations_for_image.append(point_dict)
-
-        annotations[img.pk] = annotations_for_image
-        cpc_contents[img.pk] = cpc_dict['cpc_content']
-        cpc_filenames[img.pk] = cpc_filename
-
-    if len(annotations) == 0:
-        raise FileProcessError("No matching image names found in the source")
-
-    # Note that image_dir was set the last time we picked an image match.
-    # So image_dir corresponds to an image match; that's all we want.
-    return annotations, cpc_contents, cpc_filenames, image_dir
 
 
 def annotations_preview(csv_annotations, source):
